Clean up shmem size settings

+++ csrc/executor_utils.cpp
@@ -970,7 +970,7 @@ std::vector<char> nvrtcGetCode(
 
 void dumpCompiledCodeToFile(
     const std::vector<char>& code,
-    int fusion_id,
+    int64_t fusion_id,
     bool dump_cubin) {
   std::stringstream file_name;
   file_name << "__tmp_kernel" << fusion_id << "."
@@ -1306,7 +1306,7 @@ void warnRegisterSpill(const std::string& compile_log) {
 
 void createNvrtcProgram(
     nvrtcProgram& program,
-    int id,
+    int64_t id,
     const std::string& full_src_code) {
   std::stringstream ss;
   ss << "__tmp_kernel" << id << ".cu";
@@ -1321,7 +1321,7 @@ void createNvrtcProgram(
 std::tuple<std::vector<char>, std::string> compileSource(
     const std::string& full_src_code,
     const std::string& func_name,
-    int id,
+    int64_t id,
     bool compile_to_sass,
     NvrtcCompileDriver& nvrtc_compile) {
   std::stringstream log;
@@ -1359,7 +1359,7 @@ std::tuple<NvrtcFunction, std::string, std::vector<char>> getCompiledKernel(
     c10::optional<std::reference_wrapper<const std::string>> kernel_code,
     const std::string& full_src_code,
     const std::string& func_name,
-    int id,
+    int64_t id,
     std::optional<int64_t> opt_block_size,
     const int64_t max_register_heuristic,
     bool return_compiled_binary) {
